:hammer: Refactor code
