+++ services/export/tests/unit/shared/helpers/file.test.js
@@ -15,6 +15,9 @@
  */
 const path = require('path');
 const fs = require('fs');
+
+const { interval } = require('../../../constants/csvExport');
+const { dynamicFieldsCustomDelimiter } = require('../../../constants/csv');
 const {
 	init,
 	write,
@@ -25,9 +28,16 @@ const {
 	fileExists,
 	isFile,
 	isFilePathInDirectory,
-} = require('../../../../shared/helpers/file');
 
-const { dynamicFieldsCustomDelimiter } = require('../../../constants/csv');
+	getPartialFilenameFromParams,
+	getExcelFilenameFromParams,
+	getExcelFileUrlFromParams,
+} = require('../../../../shared/helpers/file');
+const {
+	PARTIAL_FILENAME,
+	EXCEL_EXPORT_FILENAME,
+	EXCEL_FILE_URL,
+} = require('../../../../shared/regex');
 
 describe('Test filesystem interface', () => {
 	let dirPath;
@@ -123,5 +133,140 @@ describe('Test filesystem interface', () => {
 		const filePath = `${dirPath}/../../testfile.csv`;
 		const result = isFilePathInDirectory(filePath, dirPath);
 		expect(result).toBe(false);
+	});
+});
+
+describe('Test getPartialFilenameFromParams method', () => {
+	const address = 'lskpg7qukha2nmu9483huwk8oty7q3pyevh3bohr4';
+	const publicKey = '86cbecb2a176934e454f63e7ffa05783be6960d90002c5558dfd31397cd8f020';
+	const partialFilenameExtension = '.json';
+
+	it('should return partial filename when called with address', async () => {
+		const params = { address, interval: interval.startEnd };
+		const partialFilename = await getPartialFilenameFromParams(params, interval.onlyStart);
+		expect(partialFilename.endsWith(partialFilenameExtension)).toBeTruthy();
+		expect(partialFilename).toContain(address);
+		expect(partialFilename).toMatch(PARTIAL_FILENAME);
+	});
+
+	it('should return partial filename when called with publicKey', async () => {
+		const params = { publicKey, interval: interval.onlyStart };
+		const partialFilename = await getPartialFilenameFromParams(params, interval.onlyStart);
+		expect(partialFilename.endsWith(partialFilenameExtension)).toBeTruthy();
+		expect(partialFilename).toContain(address);
+		expect(partialFilename).toMatch(PARTIAL_FILENAME);
+	});
+});
+
+describe('Excel export utils', () => {
+	const address = 'lskeqretdgm6855pqnnz69ahpojk5yxfsv2am34et';
+	const publicKey = 'b7fdfc991c52ad6646159506a8326d4203c868bd3f16b8043c8e4e034346e581';
+	const excelFilenameExtension = '.xlsx';
+	const excelFileUrlBeginsWith = '/api/v3/export/';
+
+	describe('Test getExcelFilenameFromParams method', () => {
+		it('should return excel filename when called with address and complete interval with start and end date supplied', async () => {
+			const params = { address, interval: interval.startEnd };
+			const excelFilename = await getExcelFilenameFromParams(params);
+			expect(excelFilename.endsWith(excelFilenameExtension)).toBeTruthy();
+			expect(excelFilename).toContain(address);
+			expect(excelFilename).toMatch(EXCEL_EXPORT_FILENAME);
+		});
+
+		it('should return excel filename when called with publicKey and complete interval with start and end date supplied', async () => {
+			const params = { publicKey, interval: interval.startEnd };
+			const excelFilename = await getExcelFilenameFromParams(params);
+			expect(excelFilename.endsWith(excelFilenameExtension)).toBeTruthy();
+			expect(excelFilename).toContain(address);
+			expect(excelFilename).toMatch(EXCEL_EXPORT_FILENAME);
+		});
+
+		it('should return excel filename when called with address and interval with only start date supplied', async () => {
+			const params = { address, interval: interval.onlyStart };
+			const excelFilename = await getExcelFilenameFromParams(params);
+			expect(excelFilename.endsWith(excelFilenameExtension)).toBeTruthy();
+			expect(excelFilename).toContain(address);
+			expect(excelFilename).toMatch(EXCEL_EXPORT_FILENAME);
+		});
+
+		it('should return excel filename when called with publicKey and interval with only start date supplied', async () => {
+			const params = { publicKey, interval: interval.onlyStart };
+			const excelFilename = await getExcelFilenameFromParams(params);
+			expect(excelFilename.endsWith(excelFilenameExtension)).toBeTruthy();
+			expect(excelFilename).toContain(address);
+			expect(excelFilename).toMatch(EXCEL_EXPORT_FILENAME);
+		});
+
+		xit('should return excel filename when called with address and no interval supplied', async () => {
+			const params = { address };
+			const excelFilename = await getExcelFilenameFromParams(params);
+			expect(excelFilename.endsWith(excelFilenameExtension)).toBeTruthy();
+			expect(excelFilename).toContain(address);
+			expect(excelFilename).toMatch(EXCEL_EXPORT_FILENAME);
+		});
+
+		xit('should return excel filename when called with publicKey and no interval supplied', async () => {
+			const params = { publicKey };
+			const excelFilename = await getExcelFilenameFromParams(params);
+			expect(excelFilename.endsWith(excelFilenameExtension)).toBeTruthy();
+			expect(excelFilename).toContain(address);
+			expect(excelFilename).toMatch(EXCEL_EXPORT_FILENAME);
+		});
+	});
+
+	describe('Test getExcelFileUrlFromParams method', () => {
+		it('should return excel filepath URL when called with address and complete interval with start and end date supplied', async () => {
+			const params = { address, interval: interval.startEnd };
+			const excelFilepathUrl = await getExcelFileUrlFromParams(params);
+			expect(excelFilepathUrl.startsWith(excelFileUrlBeginsWith)).toBeTruthy();
+			expect(excelFilepathUrl.endsWith(excelFilenameExtension)).toBeTruthy();
+			expect(excelFilepathUrl).toContain(address);
+			expect(excelFilepathUrl).toMatch(EXCEL_FILE_URL);
+		});
+
+		it('should return excel filepath URL when called with publicKey and complete interval with start and end date supplied', async () => {
+			const params = { publicKey, interval: interval.startEnd };
+			const excelFilepathUrl = await getExcelFileUrlFromParams(params);
+			expect(excelFilepathUrl.startsWith(excelFileUrlBeginsWith)).toBeTruthy();
+			expect(excelFilepathUrl.endsWith(excelFilenameExtension)).toBeTruthy();
+			expect(excelFilepathUrl).toContain(address);
+			expect(excelFilepathUrl).toMatch(EXCEL_FILE_URL);
+		});
+
+		it('should return excel filepath URL when called with address and interval with only start date supplied', async () => {
+			const params = { address, interval: interval.onlyStart };
+			const excelFilepathUrl = await getExcelFileUrlFromParams(params);
+			expect(excelFilepathUrl.startsWith(excelFileUrlBeginsWith)).toBeTruthy();
+			expect(excelFilepathUrl.endsWith(excelFilenameExtension)).toBeTruthy();
+			expect(excelFilepathUrl).toContain(address);
+			expect(excelFilepathUrl).toMatch(EXCEL_FILE_URL);
+		});
+
+		it('should return excel filepath URL when called with publicKey and interval with only start date supplied', async () => {
+			const params = { publicKey, interval: interval.onlyStart };
+			const excelFilepathUrl = await getExcelFileUrlFromParams(params);
+			expect(excelFilepathUrl.startsWith(excelFileUrlBeginsWith)).toBeTruthy();
+			expect(excelFilepathUrl.endsWith(excelFilenameExtension)).toBeTruthy();
+			expect(excelFilepathUrl).toContain(address);
+			expect(excelFilepathUrl).toMatch(EXCEL_FILE_URL);
+		});
+
+		xit('should return excel filepath URL when called with address and no interval supplied', async () => {
+			const params = { address };
+			const excelFilepathUrl = await getExcelFileUrlFromParams(params);
+			expect(excelFilepathUrl.startsWith(excelFileUrlBeginsWith)).toBeTruthy();
+			expect(excelFilepathUrl.endsWith(excelFilenameExtension)).toBeTruthy();
+			expect(excelFilepathUrl).toContain(address);
+			expect(excelFilepathUrl).toMatch(EXCEL_FILE_URL);
+		});
+
+		xit('should return excel filepath URL when called with publicKey and no interval supplied', async () => {
+			const params = { publicKey };
+			const excelFilepathUrl = await getExcelFileUrlFromParams(params);
+			expect(excelFilepathUrl.startsWith(excelFileUrlBeginsWith)).toBeTruthy();
+			expect(excelFilepathUrl.endsWith(excelFilenameExtension)).toBeTruthy();
+			expect(excelFilepathUrl).toContain(address);
+			expect(excelFilepathUrl).toMatch(EXCEL_FILE_URL);
+		});
 	});
 });
